+++ jfreereport/source/com/jrefinery/report/io/ext/ReportConfigHandler.java
@@ -2,7 +2,7 @@
  * Date: Jan 9, 2003
  * Time: 9:08:15 PM
  *
- * $Id: ReportConfigHandler.java,v 1.2 2003/02/02 23:43:49 taqua Exp $
+ * $Id: ReportConfigHandler.java,v 1.3 2003/02/04 17:56:11 taqua Exp $
  */
 package com.jrefinery.report.io.ext;
 
@@ -163,19 +163,19 @@ public class ReportConfigHandler implements ReportDefinitionHandler
     JFreeReport report = getReport();
 
     PageFormat format = report.getDefaultPageFormat();
-    double defTopMargin = format.getImageableY();
-    double defBottomMargin = format.getHeight() - format.getImageableHeight()
-                                                - format.getImageableY();
-    double defLeftMargin = format.getImageableX();
-    double defRightMargin = format.getWidth() - format.getImageableWidth()
-                                              - format.getImageableX();
+    float defTopMargin = (float) format.getImageableY();
+    float defBottomMargin = (float) (format.getHeight() - format.getImageableHeight()
+                                                - format.getImageableY());
+    float defLeftMargin = (float) format.getImageableX();
+    float defRightMargin = (float) (format.getWidth() - format.getImageableWidth()
+                                              - format.getImageableX());
 
     format = createPageFormat(format, atts);
 
-    defTopMargin = ParserUtil.parseDouble(atts.getValue(TOPMARGIN_ATT), defTopMargin);
-    defBottomMargin = ParserUtil.parseDouble(atts.getValue(BOTTOMMARGIN_ATT), defBottomMargin);
-    defLeftMargin = ParserUtil.parseDouble(atts.getValue(LEFTMARGIN_ATT), defLeftMargin);
-    defRightMargin = ParserUtil.parseDouble(atts.getValue(RIGHTMARGIN_ATT), defRightMargin);
+    defTopMargin = ParserUtil.parseFloat(atts.getValue(TOPMARGIN_ATT), defTopMargin);
+    defBottomMargin = ParserUtil.parseFloat(atts.getValue(BOTTOMMARGIN_ATT), defBottomMargin);
+    defLeftMargin = ParserUtil.parseFloat(atts.getValue(LEFTMARGIN_ATT), defLeftMargin);
+    defRightMargin = ParserUtil.parseFloat(atts.getValue(RIGHTMARGIN_ATT), defRightMargin);
 
     Paper p = format.getPaper();
     switch (format.getOrientation())
